refactor(modules): home 7 src/services orphans into bounded contexts
src/services/ was an architectural orphan layer (flagged in
docs/ddd/EVENT_STORMING_BIG_PICTURE.md global hot spots). The 7 files
classified by their imports/consumers:

Review Execution (6 services):
- agentInstructionsBuilder.ts  (builds agent prompts for ReviewContext)
- commentLinkEnricher.ts       (file:line refs → clickable links in PR comments)
- contextActionsExecutor.ts    (executes ReviewActions via platform CLI)
- reviewContextWatcher.service.ts (watches ReviewContext file for progress)
- threadActionsExecutor.ts     (executes thread-related ReviewActions)
- threadActionsParser.ts       (parses ReviewActions from Claude output)
  → src/modules/review-execution/services/

Statistics & Insights (1 service):
- statsService.ts              (persists/loads ProjectStats and ReviewStats)
  → src/modules/statistics-insights/services/

src/services/ directory deleted (now empty).

~15 importers updated. Relative imports from src/main/ and tests
normalized to @/ alias as a side-benefit (project rule: no relative paths).

Resolves HS-4 from the DDD global hot spots.

Note: 'services/' as a layer within each module is a transitional
placeholder. Phase 2 batch B will refine each file into its proper
Clean Architecture layer (interface-adapters/, usecases helpers, or
domain services).
