+++ src/modules/review-execution/interface-adapters/gateways/cli/reviewAction.gitlab.cli.gateway.ts
@@ -1,7 +1,7 @@
 import type { ReviewAction } from '@/modules/review-execution/entities/reviewAction/reviewAction.js'
 import type { ReviewActionGateway, ExecutionContext } from '@/modules/review-execution/entities/reviewAction/reviewAction.gateway.js'
 import { ExecutionGatewayBase, type CommandInfo } from '@/shared/foundation/executionGateway.base.js'
-import { enrichCommentWithLinks } from '@/services/commentLinkEnricher.js'
+import { enrichCommentWithLinks } from '@/modules/review-execution/services/commentLinkEnricher.js'
 
 export class GitLabReviewActionCliGateway
   extends ExecutionGatewayBase<ReviewAction, ExecutionContext>
