+++ src/modules/review-execution/interface-adapters/gateways/reviewContext.fileSystem.gateway.ts
@@ -8,7 +8,7 @@ import type {
   ReviewContext,
   ReviewContextProgress,
 } from '@/modules/review-execution/entities/reviewContext/reviewContext.js'
-import { buildAgentInstructions } from '@/services/agentInstructionsBuilder.js'
+import { buildAgentInstructions } from '@/modules/review-execution/services/agentInstructionsBuilder.js'
 import type { ReviewContextAction } from '@/modules/review-execution/entities/reviewAction/reviewAction.js'
 import type { ReviewContextResult } from '@/modules/review-execution/entities/reviewContext/reviewContextResult.schema.js'
