+++ src/modules/statistics-insights/interface-adapters/controllers/http/stats.routes.ts
@@ -2,7 +2,7 @@ import type { FastifyPluginAsync } from 'fastify';
 import type { StatsGateway } from '@/modules/statistics-insights/entities/stats/stats.gateway.js';
 import type { DiffStatsFetchGateway } from '@/modules/shared-kernel/entities/diffStats/diffStatsFetch.gateway.js';
 import type { BackfillProgress } from '@/modules/statistics-insights/entities/backfill/backfillProgress.js';
-import { getStatsSummary } from '@/services/statsService.js';
+import { getStatsSummary } from '@/modules/statistics-insights/services/statsService.js';
 import { recalculateWithBackfill } from '@/modules/statistics-insights/usecases/stats/recalculateWithBackfill.usecase.js';
 import { safeParseRecalculateBody } from '@/modules/statistics-insights/entities/stats/recalculateBody.guard.js';
