+++ src/modules/tracking/interface-adapters/controllers/http/mrTrackingAdvanced.routes.ts
@@ -8,9 +8,9 @@ import { invokeClaudeReview, sendNotification } from '@/claude/invoker.js';
 import type { ReviewRequestTrackingGateway } from '../../gateways/reviewRequestTracking.gateway.js';
 import { RecordReviewCompletionUseCase } from '@/modules/tracking/usecases/tracking/recordReviewCompletion.usecase.js';
 import { SyncThreadsUseCase } from '@/modules/tracking/usecases/tracking/syncThreads.usecase.js';
-import { parseReviewOutput } from '@/services/statsService.js';
-import { parseThreadActions } from '@/services/threadActionsParser.js';
-import { executeThreadActions, defaultCommandExecutor } from '@/services/threadActionsExecutor.js';
+import { parseReviewOutput } from '@/modules/statistics-insights/services/statsService.js';
+import { parseThreadActions } from '@/modules/review-execution/services/threadActionsParser.js';
+import { executeThreadActions, defaultCommandExecutor } from '@/modules/review-execution/services/threadActionsExecutor.js';
 import { ReviewContextFileSystemGateway } from '@/modules/review-execution/interface-adapters/gateways/reviewContext.fileSystem.gateway.js';
 import { GitHubThreadFetchGateway, defaultGitHubExecutor } from '@/modules/platform-integration/interface-adapters/gateways/threadFetch.github.gateway.js';
 import { GitLabThreadFetchGateway, defaultGitLabExecutor } from '@/modules/platform-integration/interface-adapters/gateways/threadFetch.gitlab.gateway.js';
