+++ src/tests/factories/persistedInsightsData.factory.ts
@@ -1,5 +1,5 @@
 import type { PersistedDeveloperMetrics, PersistedInsightsData } from '@/modules/statistics-insights/entities/insight/persistedInsightsData.js';
-import type { ReviewStats } from '@/services/statsService.js';
+import type { ReviewStats } from '@/modules/statistics-insights/services/statsService.js';
 
 export class PersistedDeveloperMetricsFactory {
   static create(overrides: Partial<PersistedDeveloperMetrics> = {}): PersistedDeveloperMetrics {
