+++ src/tests/units/usecases/insights/computeDeveloperInsights.usecase.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { computeDeveloperInsights } from '@/modules/statistics-insights/usecases/insights/computeDeveloperInsights.usecase.js';
 import { ReviewStatsFactory } from '@/tests/factories/projectStats.factory.js';
-import type { ReviewStats } from '@/services/statsService.js';
+import type { ReviewStats } from '@/modules/statistics-insights/services/statsService.js';
 
 function createReviewsForDeveloper(
   developerName: string,
